Se creo un nuevo formato mejorado para las funciones de "removeAll"

+++ modules/removeAll.py
@@ -1,9 +1,26 @@
 import json
 import requests
+import modules.getAllData as data
 
 
 def Pago(id):
-    peticion =  requests.delete(f"http://154.38.171.54:5006/pagos/{id}")
-    res = peticion.json()
-    res["Mensaje"] = "Pago Eliminado"
-    return [res]
+    listID = []
+    for val in data.Pago():
+        listID.append(val.get("id"))
+        
+    if id in listID:
+        requests.delete(f"http://154.38.171.54:5006/pagos/{id}")
+        print("Pago eliminado")
+    else:
+        print("Error, este Pago no existe !")
+
+def Empleado(id):
+    listID = []
+    for val in data.Empleado():
+        listID.append(val.get("id"))
+        
+    if id in listID:
+        requests.delete(f"http://154.38.171.54:5006/pagos/{id}")
+        print("Empleado eliminado")
+    else:
+        print("Error, este Empleado no existe !")
